03.pyと05.pyを修正

+++ ynakamura/chapter01/05.py
@@ -6,7 +6,7 @@ fs = 16000
 sec = 3
 t = np.linspace(0., sec, int(fs * sec))
 
-whitenoise = np.random.rand(int(fs * sec)) * 2 - 1
+whitenoise = np.random.rand(fs * sec) * 2 - 1
 
 plt.plot(t, whitenoise)
 plt.xlim(0, 0.03)
